+++ demo/src/main/java/com/ringpublishing/tracking/demo/sample/SampleArticleBuilder.kt
@@ -21,7 +21,8 @@ class SampleArticleBuilder(private val resources: Resources)
 			resources.getStringArray(R.array.articles_text)[position],
 			URL(resources.getStringArray(R.array.articles_url)[position]),
 			resources.getIntArray(R.array.articles_paid)[position] > 0,
-			resources.getStringArray(R.array.articles_content_ids)[position]
+			resources.getStringArray(R.array.articles_content_ids)[position],
+			resources.getStringArray(R.array.articles_content_space_uuids)[position]
 		)
 	}
 }