+++ demo/src/main/java/com/ringpublishing/tracking/demo/sample/SamplePaidData.kt
@@ -59,5 +59,6 @@ val sampleContentMetadata = ContentMetadata(
     sourceSystemName = "source System_Name",
     contentPartIndex = 1,
     paidContent = true,
-    contentId = "my-unique-content-id-1234"
+    contentId = "my-unique-content-id-1234",
+    contentSpaceUuid = "my-unique-content-space-uuid-1234"
 )